Reorganize code in API to unburry url mapping.
